Hack around background process issues
I think, because there's no way to close stdin, that if `lh` is
started as a background process, it gets suspended.

In addition, I'm using an ugly hack to (1) use the Windows `lh.exe`
from WSL and (2) avoid aliases and treat `lh` as `lh.exe` (by
modifying the bash command-not-found handler). Realistically it's
easier to do a platform-check and just choose `lh.exe` over `lh`
as necessary.

TODO: abstract / macro for all this repeated stuff

TODO: check if also works for windows; if not, refactor to make it
cross-platform compatible

TODO: a nicer solution for quickly adjusting color + brightness